requirements.txt
pandas==2.2.3
requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.32.1